Fix ExtendedHeader fields

+++ dltparser/src/commonMain/kotlin/com/alekso/dltparser/dlt/ExtendedHeader.kt
@@ -3,7 +3,7 @@ package com.alekso.dltparser.dlt
 
 data class ExtendedHeader(
     val messageInfo: MessageInfo,
-    val argumentsCount: Int,
+    val argumentsCount: UByte,
     val applicationId: String,
     val contextId: String,
 ) {
